// src/src2/api/models/LigneProformaResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LigneProformaResponse = {
    idProduit?: string;
    nomProduit?: string;
    description?: string;
    quantite?: number;
    prixUnitaire?: number;
    montantTotal?: number;
    remisePourcentage?: number;
    remiseMontant?: number;
    tauxTva?: number;
};